% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtables.R
\name{addtablesUI}
\alias{addtablesUI}
\title{addtablesUI}
\usage{
addtablesUI(id, M)
}
\arguments{
\item{id}{is caller id}

\item{M}{is the meta data connection structure}
}
\description{
UI for adding tables to 'cuborg' data warehouse
}
