% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsource.R
\name{cfgdataUI}
\alias{cfgdataUI}
\title{cfgdataUI}
\usage{
cfgdataUI(id, M)
}
\arguments{
\item{id}{is caller id}

\item{M}{is the meta data connection structure}
}
\description{
UI for configuring new data source for ingestion
}
